requirements.txt
flask==3.1.2
python-dotenv==1.1.1
google-generativeai==0.8.5
